requirements.txt
Django==4.2.8
Pillow==10.0.1
reportlab==4.0.8